# api-gateway/src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZyvoPay API Gateway",
    "description": "ZyvoPay Payment Gateway API - Complete banking-as-a-service platform for PIX payments, TED transfers, and financial operations.\n\n## Authentication\nAll endpoints (except `/health` and `/api/auth/clients`) require API Key authentication.\n\nProvide your API key via:\n- `X-API-Key` header, or\n- `Authorization: Bearer <api_key>` header\n\n## Rate Limiting\n- Default: 60 requests per minute per API key\n- Rate limit headers are returned in responses\n\n## Webhooks\nConfigure your webhook URL in the client settings to receive real-time event notifications.\n",
    "version": "1.0.0",
    "contact": {
      "name": "ZyvoPay Support",
      "email": "support@zyvopay.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.zyvopay.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/pix/keys": {
      "post": {
        "summary": "Register PIX key",
        "tags": [
          "PIX"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "subaccount_id",
                  "key_type",
                  "key_value"
                ],
                "properties": {
                  "subaccount_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "key_type": {
                    "type": "string",
                    "enum": [
                      "CPF",
                      "CNPJ",
                      "EMAIL",
                      "PHONE",
                      "EVP"
                    ]
                  },
                  "key_value": {
                    "type": "string"
                  },
                  "is_default": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PIX key registered"
          }
        }
      },
      "get": {
        "summary": "List PIX keys",
        "tags": [
          "PIX"
        ],
        "parameters": [
          {
            "name": "subaccount_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of PIX keys"
          }
        }
      }
    },
    "/api/pix/keys/{id}": {
      "delete": {
        "summary": "Delete PIX key",
        "tags": [
          "PIX"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PIX key deleted"
          }
        }
      }
    },
    "/api/pix/payments": {
      "post": {
        "summary": "Create PIX payment",
        "tags": [
          "PIX"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "subaccount_id",
                  "pix_key",
                  "pix_key_type",
                  "amount_cents"
                ],
                "properties": {
                  "subaccount_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "pix_key": {
                    "type": "string"
                  },
                  "pix_key_type": {
                    "type": "string",
                    "enum": [
                      "CPF",
                      "CNPJ",
                      "EMAIL",
                      "PHONE",
                      "EVP"
                    ]
                  },
                  "amount_cents": {
                    "type": "integer"
                  },
                  "description": {
                    "type": "string"
                  },
                  "client_request_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PIX payment created"
          }
        }
      }
    },
    "/api/pix/payments/{id}": {
      "get": {
        "summary": "Get PIX payment status",
        "tags": [
          "PIX"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment status"
          }
        }
      }
    },
    "/api/pix/qrcodes": {
      "post": {
        "summary": "Generate PIX QR Code",
        "tags": [
          "PIX"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "subaccount_id",
                  "qr_type"
                ],
                "properties": {
                  "subaccount_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "qr_type": {
                    "type": "string",
                    "enum": [
                      "static",
                      "dynamic"
                    ]
                  },
                  "amount_cents": {
                    "type": "integer"
                  },
                  "description": {
                    "type": "string"
                  },
                  "expires_in_minutes": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "QR code generated"
          }
        }
      },
      "get": {
        "summary": "List QR codes",
        "tags": [
          "PIX"
        ],
        "parameters": [
          {
            "name": "subaccount_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "qr_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of QR codes"
          }
        }
      }
    },
    "/api/pix/qrcodes/{id}": {
      "get": {
        "summary": "Get QR code details",
        "tags": [
          "PIX"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QR code details"
          }
        }
      }
    },
...
  }
}